# docs/api/openapi.yaml
openapi: '3.1.0'
info:
  title: Orchestrate OS API
  version: 2.5.0-draft
  description: 'Governance-only semantic control plane API. Resource-style endpoints for multi-user Postgres-backed persistence
    with append-only audit logging.

    '
  contact:
    name: Orchestrate OS Team
servers:
- url: /api/v2.5
  description: Canonical API base
paths:
...
  /documents/{doc_id}/reader-nodes:
    get:
      operationId: getReaderNodes
      summary: Get cached reader nodes for a document
      description: 'Returns cached text extraction nodes for a document. Cache key is (document_id, source_pdf_hash, ocr_version).
        If no cache entry exists, returns empty nodes with quality_flag=missing_text_layer. Feature-gated behind EVIDENCE_INSPECTOR_V251.

        '
      tags:
      - Evidence Inspector
      parameters:
      - name: doc_id
        in: path
        required: true
        schema:
          type: string
      - name: source_pdf_hash
        in: query
        schema:
          type: string
        description: SHA-256 hash of the source PDF
      - name: ocr_version
        in: query
        schema:
          type: string
          default: v1
      responses:
        '200':
          description: Reader node cache hit or fallback
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReaderNodeResponse'
        '404':
          description: Document not found or feature disabled
    put:
      operationId: upsertReaderNodes
      summary: Upsert reader node cache for a document
      description: 'Creates or updates the reader node cache entry for a document. Each node receives a deterministic node_id
        based on (document_id, page_number, block_index). Quality flag is auto-detected if not provided. Cache key: (document_id,
        source_pdf_hash, ocr_version). Emits reader_node_cache.upserted audit event. Feature-gated behind EVIDENCE_INSPECTOR_V251.'
      tags:
      - Evidence Inspector
      parameters:
      - name: doc_id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - source_pdf_hash
              - nodes
              properties:
                source_pdf_hash:
                  type: string
                  description: SHA-256 hash of the source PDF bytes
                ocr_version:
                  type: string
                  default: v1
                  description: OCR engine version string
                page_count:
                  type: integer
                  description: Total page count in the PDF
                quality_flag:
                  type: string
                  enum:
                  - ok
                  - suspect_mojibake
                  - unreadable
                  - missing_text_layer
                  description: Override quality flag (auto-detected if omitted)
                nodes:
                  type: array
                  items:
                    type: object
                    properties:
                      page_number:
                        type: integer
                      block_index:
                        type: integer
                      text:
                        type: string
                      bbox:
                        type: array
                        items:
                          type: number
                        description: '[x0, y0, x1, y1] bounding box'
      responses:
        '200':
          description: Reader node cache created or updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReaderNodeResponse'
        '400':
          description: Validation error (missing source_pdf_hash)
        '404':
          description: Document not found or feature disabled
...
components:
...
  schemas:
...
    ResponseMeta:
      type: object
      properties:
        request_id:
          type: string
        timestamp:
          type: string
          format: date-time
...
    ReaderNodeCache:
      type: object
      properties:
        id:
          type: string
        document_id:
          type: string
        source_pdf_hash:
          type: string
        ocr_version:
          type: string
        quality_flag:
          type: string
          enum:
          - ok
          - suspect_mojibake
          - unreadable
          - missing_text_layer
        nodes:
          type: array
          items:
            type: object
        page_count:
          type: integer
        cached:
          type: boolean
          description: 'false when returning fallback (no cache entry found), true or omitted when returning from DB cache

            '
        created_at:
          type: string
          format: date-time
        metadata:
          type: object
    ReaderNodeResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/ReaderNodeCache'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
...
  securitySchemes:
    GoogleOAuth:
      type: oauth2
      description: 'Google OAuth 2.0 (OIDC) for production human users. Required for human-governed endpoints (patch transitions,
        approvals, annotations).

        '
      flows:
        authorizationCode:
          authorizationUrl: https://accounts.google.com/o/oauth2/v2/auth
          tokenUrl: https://oauth2.googleapis.com/token
          scopes:
            email: Read user email
            profile: Read user profile
    BearerToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'Server-issued JWT (1h expiry) obtained after Google OAuth code exchange. Used in Authorization header
        for all authenticated requests.

        '
    ScopedApiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: 'Workspace-scoped API key for sandbox and service ingestion endpoints. Keys carry explicit scopes (e.g.,
        signals:write, batches:write, read:all). Required for service ingestion endpoints (POST /signals, POST /triage-items).
        Accepted as alternative auth on read endpoints.

        '
security:
- BearerToken: []
- ScopedApiKey: []